backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.4.1
torchvision==0.19.1
pillow==10.4.0
numpy==1.26.4
requests==2.31.0